+++ packages/dai-plugin-migrations/package.json
@@ -3,9 +3,11 @@
   "description": "Plugin to add migrations to dai.js",
   "version": "1.3.17",
   "license": "MIT",
-  "main": "dist/src/index.js",
-  "unpkg": "dist/src/index.js",
-  "jsdelivr": "dist/src/index.js",
+  "main": "dist/index.js",
+  "unpkg": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "module": "dist/dai-plugin-migrations.esm.js",
+  "jsdelivr": "dist/dai-plugin-migrations.esm.js",
   "scripts": {
     "build": "yarn clean  && tsdx build",
     "clean": "rm -rf ./dist",
