+++ packages/dai-plugin-migrations/src/tokens.ts
@@ -1,0 +1,8 @@
+import { createCurrency } from '@makerdao/currency';
+
+export const OLD_MKR = createCurrency('OLD_MKR');
+export const SAI = createCurrency('SAI');
+export const DAI = createCurrency('DAI');
+export const MKR = createCurrency('MKR');
+
+export const DAI_1 = createCurrency('DAI_1');
